requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
asgiref==3.8.1
    # via
    #   -r /home/mgmdy/Work/Envs/Local/nightly/requirements/openedx-filters-samples/requirements/base.txt
    #   django
click==8.1.8
    # via code-annotations
code-annotations==2.2.0
    # via -r requirements/test.in
coverage[toml]==7.6.10
    # via pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r /home/mgmdy/Work/Envs/Local/nightly/requirements/openedx-filters-samples/requirements/base.txt
    #   openedx-filters
dnspython==2.7.0
    # via
    #   -r /home/mgmdy/Work/Envs/Local/nightly/requirements/openedx-filters-samples/requirements/base.txt
    #   pymongo
edx-opaque-keys[django]==2.11.0
    # via -r /home/mgmdy/Work/Envs/Local/nightly/requirements/openedx-filters-samples/requirements/base.txt
iniconfig==2.0.0
    # via pytest
jinja2==3.1.5
    # via code-annotations
markupsafe==3.0.2
    # via jinja2
openedx-filters==1.12.0
    # via -r /home/mgmdy/Work/Envs/Local/nightly/requirements/openedx-filters-samples/requirements/base.txt
packaging==24.2
    # via pytest
pbr==6.1.0
    # via
    #   -r /home/mgmdy/Work/Envs/Local/nightly/requirements/openedx-filters-samples/requirements/base.txt
    #   stevedore
pluggy==1.5.0
    # via pytest
pymongo==4.10.1
    # via
    #   -r /home/mgmdy/Work/Envs/Local/nightly/requirements/openedx-filters-samples/requirements/base.txt
    #   edx-opaque-keys
pytest==8.3.4
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==6.0.0
    # via -r requirements/test.in
pytest-django==4.9.0
    # via -r requirements/test.in
python-slugify==8.0.4
    # via code-annotations
pyyaml==6.0.2
    # via code-annotations
sqlparse==0.5.3
    # via
    #   -r /home/mgmdy/Work/Envs/Local/nightly/requirements/openedx-filters-samples/requirements/base.txt
    #   django
stevedore==5.4.0
    # via
    #   -r /home/mgmdy/Work/Envs/Local/nightly/requirements/openedx-filters-samples/requirements/base.txt
    #   code-annotations
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.12.2
    # via
    #   -r /home/mgmdy/Work/Envs/Local/nightly/requirements/openedx-filters-samples/requirements/base.txt
    #   edx-opaque-keys
